DATAMONGO-1135 - Convert GeoJson to $geometry.
We now support GeoJson structures for queries. Depending on the GeoSpatialIndexType we use the legacy format for 2D and GeoJson for 2D_Sphere indexes. Therfore we inspect the query for specific geo spatial keywords and convert the query parameter according to the information we read from the Entities meta model.